Update supplementary-information-brief.Rmd
I was able to replicate the issue with Talker.SpeechDescription, but this seems limited to publishing with kable(). The raw dataframe has the correct number of rows and doesn't have extra comments hanging in space. For now, I've removed the kable() function (after trying a few unsuccesful fixes such as removing numbers, special characters, or setting escape = T). If it's important to have that table be nicely formatted we can revisit and bugfix!